refactor: split AppBar into a couple ot the components
